--- tplink-energy-monitor/apps/client/src/app/models/devices/device-type.enum.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line no-shadow
-export enum TpLinkDeviceTypes {
-    SMARTPLUGSWITCH = 'IOT.SMARTPLUGSWITCH',
-    RANGEEXTENDER = 'IOT.RANGEEXTENDER.SMARTPLUG',
-}
