--- tplink-energy-monitor/apps/client/src/app/models/devices/tp-link-plug.dto.ts
@@ -1,18 +1,0 @@
-import { TpLinkDeviceTypes } from './device-type.enum';
-import { DeviceEnergyOverview, DeviceMonthlyEnergyOverview } from './tp-link-plug-info.dto';
-
-export type TPLinkPlug = {
-    id: string;
-    name: string;
-    alias: string;
-    type: TpLinkDeviceTypes;
-    model: string;
-    deviceName: string;
-    relayState: boolean;
-    hwId: string;
-    host: string;
-    swVersion: string;
-    hwVersion: string;
-    last30Days: DeviceEnergyOverview;
-    last12Month: DeviceMonthlyEnergyOverview;
-};
