--- tplink-energy-monitor/apps/client/src/app/models/settings/user-settings.dto.ts
@@ -1,4 +1,0 @@
-export type UserSettings = {
-    energyCosts: number;
-    currency: 'EUR' | 'USD';
-};
