+++ tplink-energy-monitor/apps/client/src/app/redux/devices/actions/update-all-devices.action.ts
@@ -1,5 +1,5 @@
 import { DevicesAction, DevicesActionNames } from '../devices-state.type';
-import { TPLinkPlug } from '../../../models/devices/tp-link-plug.dto';
+import { TPLinkPlug } from '@tplink-energy-monitor/data-access-devices';
 
 export const updateAllDevices = (devices: TPLinkPlug[]): DevicesAction<TPLinkPlug[]> => {
     return {
