+++ tplink-energy-monitor/jest.config.js
@@ -1,3 +1,8 @@
 module.exports = {
-  projects: ['<rootDir>/apps/api-server', '<rootDir>/apps/client'],
+  projects: [
+    '<rootDir>/apps/api-server',
+    '<rootDir>/apps/client',
+    '<rootDir>/libs/data-access-user-settings',
+    '<rootDir>/libs/data-access-devices',
+  ],
 };
