+++ tplink-energy-monitor/libs/data-access-devices/src/lib/device-type.enum.ts
