+++ tplink-energy-monitor/tsconfig.base.json
@@ -16,7 +16,14 @@
     "skipLibCheck": true,
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
-    "paths": {}
+    "paths": {
+      "@tplink-energy-monitor/data-access-user-settings": [
+        "libs/data-access-user-settings/src/index.ts"
+      ],
+      "@tplink-energy-monitor/data-access-devices": [
+        "libs/data-access-devices/src/index.ts"
+      ]
+    }
   },
   "exclude": ["node_modules", "tmp"]
 }
